backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.0
python-dotenv==1.0.0